src/commands/LinkCommand.js -> commands/link/index.js
