+++ commands/link/__tests__/__fixtures__/basic/packages/package-2/cli.js
