+++ commands/link/__tests__/__fixtures__/basic/packages/package-3/cli2.js
