+++ commands/link/__tests__/__fixtures__/force-local/packages/package-2/cli.js
