+++ commands/link/__tests__/__fixtures__/force-local/packages/package-3/cli1.js
