+++ commands/link/__tests__/__fixtures__/force-local/packages/package-4/package.json
